moved common bulk storage methods to abstract class
 - removed one of the set storage methods because it was not used
