+++ modMcf/src/org/aion/mcf/db/AbstractContractDetails.java
@@ -4,6 +4,8 @@ import static org.aion.crypto.HashUtil.EMPTY_DATA_HASH;
 import static org.aion.crypto.HashUtil.h256;
 import static org.aion.util.bytes.ByteUtil.EMPTY_BYTE_ARRAY;
 
+import com.google.common.annotations.VisibleForTesting;
+import java.util.Collection;
 import java.util.HashMap;
 import java.util.Map;
 import org.aion.base.db.IContractDetails;
@@ -100,5 +102,41 @@ public abstract class AbstractContractDetails implements IContractDetails {
                         + "\n";
         ret += "  Storage: " + getStorageHash();
         return ret;
+    }
+
+    @VisibleForTesting
+    @Override
+    public void setStorage(Map<ByteArrayWrapper, ByteArrayWrapper> storage) {
+        for (Map.Entry<ByteArrayWrapper, ByteArrayWrapper> entry : storage.entrySet()) {
+            ByteArrayWrapper key = entry.getKey();
+            ByteArrayWrapper value = entry.getValue();
+
+            if (value != null) {
+                put(key, value);
+            } else {
+                delete(key);
+            }
+        }
+    }
+
+    @Override
+    public Map<ByteArrayWrapper, ByteArrayWrapper> getStorage(Collection<ByteArrayWrapper> keys) {
+        Map<ByteArrayWrapper, ByteArrayWrapper> storage = new HashMap<>();
+
+        if (keys == null) {
+            throw new IllegalArgumentException("Input keys cannot be null");
+        } else {
+            for (ByteArrayWrapper key : keys) {
+                ByteArrayWrapper value = get(key);
+
+                // we check if the value is not null,
+                // cause we keep all historical keys
+                if ((value != null) && (!value.isZero())) {
+                    storage.put(key, value);
+                }
+            }
+        }
+
+        return storage;
     }
 }
